Fix two thinkos.

+++ source_distiller.py
@@ -119,7 +119,8 @@ class MergeModeDistiller(SourceDistiller):
                 files = glob.glob(tempdir+'/*')
                 # If everything is in a single dir then move everything up one
                 # level.
-                shutil.rmtree(target)
+                if os.path.isdir(target):
+                    shutil.rmtree(target)
                 if len(files) == 1:
                     shutil.move(files[0], target)
                 else:
